refactor: resolve all clippy too_many_arguments warnings
Pack business parameters into typed structs (CreateExecutionParams,
UpdateStatusParams, CreateNotificationParams, SysSmallUploadParams,
PipelineParams, CreateMessageParams, CreateDocumentParams,
HybridSearchParams, IndexMessageParams, VectorRecallParams,
RecallHistoryParams, CompactHistoryParams, SpawnIndexParams,
QaStreamParams) while keeping db/ctx as standalone infra deps.

Also fixes: needless_range_loop, clamp, doc formatting, match arm
merges, redundant closures, uninlined format args, use_self,
map_or -> is_none_or, and other minor clippy warnings.

cargo clippy --all-features -- -D warnings passes clean.

+++ src/modules/knowledge_base/service/chunking_service.rs
@@ -152,7 +152,7 @@ fn flush_chunk(
 ) {
     let char_start = byte_to_char(markdown, byte_start) as i32;
     let char_end = char_start + content.chars().count() as i32;
-    let token_count = count_tokens(content) as i32;
+    let token_count = count_tokens(content);
     let heading_path = resolve_heading_path(byte_start, heading_map);
 
     chunks.push(RawChunk {
